Feature: Adicionado logica para o dash de Logistica

+++ routes/movimentos.py
@@ -10,6 +10,13 @@ def movimentos_root():
         case "GET": return movement_service.read()
         case "POST": return movement_service.create()
         
-@movimentos_bp.route("/<int:id>", methods=["DELETE"])
+@movimentos_bp.get("/dashboard")
+def movements_dashboard():
+    return movement_service.dashboard()
+
+
+@movimentos_bp.route("/<int:id>", methods=["PATCH", "DELETE"])
 def movimentos_by_id(id):
+    if request.method == "PATCH":
+        return movement_service.update(id)
     return movement_service.delete(id)
